database tables set up and migrated

+++ database/migrations/2023_10_25_104105_create_contacts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->timestamps();
             $table->string('name');
-            $table->long('userID');
+            $table->integer('userID');
             $table->string('email');
             $table->string('message');
         });
